Refactor InputResult out of AbstractSpoofaxResult hierarchy
