+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/core/IEvaluationStrategy.java
@@ -2,8 +2,6 @@ package org.metaborg.spoofax.shell.core;
 
 import org.metaborg.core.MetaborgException;
 import org.metaborg.core.context.IContext;
-import org.metaborg.spoofax.shell.output.AnalyzeResult;
-import org.metaborg.spoofax.shell.output.ParseResult;
 import org.spoofax.interpreter.terms.IStrategoTerm;
 
 /**
@@ -18,19 +16,11 @@ public interface IEvaluationStrategy {
 
     /**
      * Evaluate the given Stratego term using this strategy.
-     * @param parsed The parsed input.
+     * @param term The input term.
      * @param context The {@link IContext}.
      * @return The output Stratego term.
      * @throws MetaborgException When evaluation fails for one reason or another.
      */
-    IStrategoTerm evaluate(ParseResult parsed, IContext context) throws MetaborgException;
+    IStrategoTerm evaluate(IStrategoTerm term, IContext context) throws MetaborgException;
 
-    /**
-     * Evaluate the given Stratego term using this strategy.
-     * @param analyzed The analyzed input.
-     * @param context The {@link IContext}.
-     * @return The output Stratego term.
-     * @throws MetaborgException When evaluation fails for one reason or another.
-     */
-    IStrategoTerm evaluate(AnalyzeResult analyzed, IContext context) throws MetaborgException;
 }
