+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/functions/EvaluateFunction.java
@@ -13,20 +13,21 @@ import org.metaborg.spoofax.shell.output.EvaluateResult;
 import org.metaborg.spoofax.shell.output.FailOrSuccessResult;
 import org.metaborg.spoofax.shell.output.IResult;
 import org.metaborg.spoofax.shell.output.IResultFactory;
-import org.metaborg.spoofax.shell.output.ParseResult;
+import org.metaborg.spoofax.shell.output.ISpoofaxTermResult;
 import org.spoofax.interpreter.terms.IStrategoTerm;
 
 import com.google.inject.Inject;
 import com.google.inject.assistedinject.Assisted;
 
 /**
- * Creates an {@link EvaluateResult} from a given {@link ParseResult}.
+ * Creates an {@link EvaluateResult} from a given {@link AnalyzeResult}.
  */
-public class PEvalFunction extends ContextualSpoofaxFunction<ParseResult, EvaluateResult> {
+public class EvaluateFunction extends ContextualSpoofaxFunction<ISpoofaxTermResult<?>,
+                                                                EvaluateResult> {
     private final Map<String, IEvaluationStrategy> evaluationStrategies;
 
     /**
-     * Instantiate a {@link PEvalFunction}.
+     * Instantiate an {@link AEvalFunction}.
      *
      * @param evaluationStrategies
      *            The {@link IEvaluationStrategy} implementations, grouped by their names as keys.
@@ -40,7 +41,7 @@ public class PEvalFunction extends ContextualSpoofaxFunction<ParseResult, Evalua
      *            The {@link ILanguageImpl} to which this command applies.
      */
     @Inject
-    public PEvalFunction(Map<String, IEvaluationStrategy> evaluationStrategies,
+    public EvaluateFunction(Map<String, IEvaluationStrategy> evaluationStrategies,
                          IContextService contextService, IResultFactory resultFactory,
                          @Assisted IProject project, @Assisted ILanguageImpl lang) {
         super(contextService, resultFactory, project, lang);
@@ -49,14 +50,18 @@ public class PEvalFunction extends ContextualSpoofaxFunction<ParseResult, Evalua
 
     @Override
     protected FailOrSuccessResult<EvaluateResult, IResult>
-            applyThrowing(IContext context, ParseResult a) throws Exception {
+            applyThrowing(IContext context, ISpoofaxTermResult<?> a) throws Exception {
+        if (!a.ast().isPresent()) {
+            return FailOrSuccessResult.failed(a);
+        }
         ShellFacet facet = context.language().facet(ShellFacet.class);
         if (facet == null) {
             throw new MetaborgException("Cannot find the shell facet.");
         }
 
         IEvaluationStrategy evalStrategy = evaluationStrategies.get(facet.getEvaluationMethod());
-        IStrategoTerm result = evalStrategy.evaluate(a, context);
+        IStrategoTerm result = evalStrategy.evaluate(a.ast().get(), context);
+
         return FailOrSuccessResult.ofSpoofaxResult(resultFactory.createEvaluateResult(a, result));
     }
 }
