+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/output/EvaluateResult.java
@@ -7,8 +7,6 @@ import org.metaborg.core.context.IContext;
 import org.metaborg.core.messages.IMessage;
 import org.metaborg.core.unit.IUnit;
 import org.metaborg.spoofax.core.stratego.IStrategoCommon;
-import org.metaborg.spoofax.shell.functions.AEvalFunction;
-import org.metaborg.spoofax.shell.functions.PEvalFunction;
 import org.spoofax.interpreter.terms.IStrategoTerm;
 
 import com.google.inject.Inject;
@@ -17,57 +15,24 @@ import com.google.inject.assistedinject.Assisted;
 /**
  * The result of the execution of an {@link AEvalFunction} or a {@link PEvalFunction}.
  */
-public abstract class EvaluateResult extends AbstractSpoofaxResult<IUnit> {
-    private final AbstractSpoofaxResult<?> wrappedDelegate;
+public class EvaluateResult extends AbstractSpoofaxTermResult<IUnit> {
+    private final ISpoofaxTermResult<?> wrappedDelegate;
     private final IStrategoTerm result;
 
     /**
-     * The result of the evaluation of an analyzed AST.
+     * Create a {@link EvaluateResult}.
+     *
+     * @param common
+     *            the {@link IStrategoCommon} service.
+     * @param wrappedResult
+     *            the wrapped {@link ISpoofaxTermResult}.
+     * @param result
+     *            the result of the evaluation.
      */
-    public static class Analyzed extends EvaluateResult {
-
-        /**
-         * Create a {@link EvaluateResult}.
-         *
-         * @param common
-         *            the {@link IStrategoCommon} service
-         * @param analyzed
-         *            the wrapped {@link AnalyzeResult}.
-         * @param result
-         *            the result of the evaluation.
-         */
-        @Inject
-        public Analyzed(IStrategoCommon common, @Assisted AnalyzeResult analyzed,
-                        @Assisted IStrategoTerm result) {
-            super(common, analyzed, result);
-        }
-    }
-
-    /**
-     * The result of the evaluation of a parsed, but not analyzed, AST.
-     */
-    public static class Parsed extends EvaluateResult {
-
-        /**
-         * Create a {@link EvaluateResult}.
-         *
-         * @param common
-         *            the {@link IStrategoCommon} service
-         * @param parsed
-         *            the wrapped {@link ParseResult}.
-         * @param result
-         *            the result of the evaluation.
-         */
-        @Inject
-        public Parsed(IStrategoCommon common, @Assisted ParseResult parsed,
-                      @Assisted IStrategoTerm result) {
-            super(common, parsed, result);
-        }
-    }
-
-    private <T extends IUnit> EvaluateResult(IStrategoCommon common,
-                                             AbstractSpoofaxResult<T> wrappedResult,
-                                             IStrategoTerm result) {
+    @Inject
+    public EvaluateResult(IStrategoCommon common,
+                          @Assisted ISpoofaxTermResult<?> wrappedResult,
+                          @Assisted IStrategoTerm result) {
         super(common, wrappedResult.unit());
         this.wrappedDelegate = wrappedResult;
         this.result = result;
