+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/output/ITermResult.java
@@ -1,0 +1,17 @@
+package org.metaborg.spoofax.shell.output;
+
+import java.util.Optional;
+
+import org.spoofax.interpreter.terms.IStrategoTerm;
+
+/**
+ * Interface for results that may have an AST.
+ */
+public interface ITermResult {
+
+    /**
+     * Returns the ast of this unit as a {@link IStrategoTerm} if present.
+     * @return a {@link IStrategoTerm} or null
+     */
+    Optional<IStrategoTerm> ast();
+}
