+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/output/ParseResult.java
@@ -19,7 +19,7 @@ import com.google.inject.assistedinject.AssistedInject;
  * Represents a {@link ParseResult} as returned by the {@link IReplCommand}.
  * Wraps a {@link ISpoofaxParseUnit}.
  */
-public class ParseResult extends AbstractSpoofaxResult<ISpoofaxParseUnit> {
+public class ParseResult extends AbstractSpoofaxTermResult<ISpoofaxParseUnit> {
 
     /**
      * Create a {@link ParseResult}.
