+++ org.metaborg.spoofax.shell.core/src/main/java/org/metaborg/spoofax/shell/output/TransformResult.java
@@ -24,7 +24,7 @@ import com.google.inject.assistedinject.AssistedInject;
  * {@link ISpoofaxTransformUnit}.
  */
 public abstract class TransformResult
-    extends AbstractSpoofaxResult<ISpoofaxTransformUnit<?>> {
+    extends AbstractSpoofaxTermResult<ISpoofaxTransformUnit<?>> {
 
     /**
      * The result of the transformation of an analyzed AST.
